+++ source/extensions/lightspeed.trex.control.stagecraft/lightspeed/trex/control/stagecraft/tests/__init__.py
@@ -15,17 +15,17 @@
 * limitations under the License.
 """
 
+from .e2e.test_capture_repair_on_open import TestCaptureRepairOnOpen
 from .e2e.test_capture_swap_undo import TestCaptureSwapUndo
 from .e2e.test_hotkeys import TestHotkeys
 from .e2e.test_stage_manager import TestStageManagerPropertiesInteraction
-from .unit.test_check_capture_on_open import TestCheckCaptureOnOpen
 from .unit.test_commands import TestCommands
 from .unit.test_setup import TestSetup as TestSetupUnit
 from .unit.test_stage_manager_schema import TestStageManagerSchema
 
 __all__ = [
+    "TestCaptureRepairOnOpen",
     "TestCaptureSwapUndo",
-    "TestCheckCaptureOnOpen",
     "TestCommands",
     "TestHotkeys",
     "TestSetupUnit",