+++ source/extensions/lightspeed.trex.project_wizard.open_project_page.widget/lightspeed/trex/project_wizard/open_project_page/widget/tests/e2e/__init__.py
@@ -15,11 +15,6 @@
 * limitations under the License.
 """
 
-from .e2e.test_home_widget import TestHomeWidgetInvalidDepsFlow
-from .unit.test_home_widget import TestLoadWorkFile, TestRefreshRecentItemsDeferred
+from .test_setup_ui import TestWizardOpenProjectPage
 
-__all__ = [
-    "TestHomeWidgetInvalidDepsFlow",
-    "TestLoadWorkFile",
-    "TestRefreshRecentItemsDeferred",
-]
+__all__ = ["TestWizardOpenProjectPage"]